refactor: split tests in separated modules
